Add start pos dependent playback after navigation.
Creates a filename based subclass of RecordedTeleop so that multiple
programs can easily be recorded. Based on start position, plays back a
corresponding file after navigating to the correct column for the sake
of scoring the glyph.

In order for future steps to be added, the playback code needs to be
updated to return when playback has completed.